# swagger.yaml
openapi: 3.0.0
info:
  title: Unofficial UMDearborn Catalog API
  version: 0.0.0
components:
  schemas:
    term:
      properties:
        code:
          type: string
        season:
          type: string
        year:
          type: number
    subject:
      properties:
        code:
          type: string
        name:
          type: string
paths:
  '/api/catalog/terms':
    get:
      summary: Returns a list of terms and their codes.
      responses:
        '200':
          description: A JSON array of user names
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/term'
  '/api/catalog/subjects/{termCode}':
    get:
      parameters:
        - in: path
          name: termCode
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Returns a list of subjects in a term
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/subject'
  '/api/catalog/entries/{termCode}/{subjectCode}':
    get:
      parameters:
        - in: path
          name: termCode
          required: true
          schema:
            type: string
        - in: path
          name: subjectCode
          required: true
          schema:
            type: string
      responses:
        '200':
          description: >-
            Returns a list of catalog entries available during the given term in
            the given subject
          content:
            application/json:
              schema:
                type: array
  '/api/catalog/course-details/{termCode}/{subjectCode}/{courseNumber}':
    get:
      parameters:
      - in: path
        name: termCode
        required: true
        schema:
          type: string
      - in: path
        name: subjectCode
        required: true
        schema:
          type: string
      - in: path
        name: courseNumber
        required: true
        schema:
          type: string
      responses:
        '200':
          description: >-
            Returns the course description and its prerequisties.
          content:
            application/json:
              schema:
                type: array
  '/api/catalog/schedule-listings/{termCode}/{subjectCode}/{courseNumber}/{scheduleTypeCode}':
    get:
      parameters:
      - in: path
        name: termCode
        required: true
        schema:
          type: string
      - in: path
        name: subjectCode
        required: true
        schema:
          type: string
      - in: path
        name: courseNumber
        required: true
        schema:
          type: string
      - in: path
        name: scheduleTypeCode
        required: true
        schema:
          type: string
      responses:
        '200':
          description: >-
            Returns a list of course registration numbers (CRNs).
          content:
            application/json:
              schema:
                type: array
  '/api/catalog/schedule-detail/{termCode}/{crn}':
    get:
      parameters:
      - in: path
        name: termCode
        required: true
        schema:
          type: string
      - in: path
        name: crn
        required: true
        schema:
          type: string
      responses:
        '200':
          description: >-
            Returns a list of course registration numbers (CRNs).
          content:
            application/json:
              schema:
                type: array